confirma existência do diretório antes de deletar

+++ Janelas/CheckerPopup.xaml.cs
@@ -319,11 +319,14 @@ namespace TrueMiningDesktop.Janelas
 							catch { };
 						}
 					}
+
+					//temp: remove pasta antiga de bibliotecas, que deixou de ser usada (foi substituída por outra pasta) a partir da versão 4.20
+					if (Directory.Exists("libraries"))
+					{
+						Directory.Delete("libraries", true);
+					}
 				}
 				catch { };
-
-				//temp: remove pasta antiga de bibliotecas, que deixou de ser usada (foi substituída por outra pasta) a partir da versão 4.20
-				Directory.Delete("libraries", true);
 			});
 			removeOldFiles.Start();
 			removeOldFiles.Wait(4000);
